+++ src/plugins/pouchdb/pouchdb-helper.ts
@@ -7,6 +7,7 @@ import type {
     RxAttachmentWriteData,
     RxDocumentData,
     RxDocumentWriteData,
+    RxLocalDocumentData,
     WithAttachments
 } from '../../types';
 import type { RxStorageInstancePouch } from './rx-storage-instance-pouch';
@@ -35,6 +36,8 @@ export const OPEN_POUCHDB_STORAGE_INSTANCES: Set<RxStorageKeyObjectInstancePouch
  * prefix of local pouchdb documents
  */
 export const POUCHDB_LOCAL_PREFIX: '_local/' = '_local/';
+export const POUCHDB_LOCAL_PREFIX_LENGTH = POUCHDB_LOCAL_PREFIX.length;
+
 /**
  * Pouchdb stores indexes as design documents,
  * we have to filter them out and not return the
@@ -345,4 +348,50 @@ export function getPouchIndexDesignDocNameByIndex(
 ): string {
     const indexName = 'idx-rxdb-index-' + index.join(',');
     return indexName;
+}
+
+/**
+ * PouchDB has not way to read deleted local documents
+ * out of the database.
+ * So instead of deleting them, we set a custom deleted flag.
+ */
+export const RXDB_POUCH_DELETED_FLAG = 'rxdb-pouch-deleted' as const;
+
+
+export type RxLocalDocumentDataWithCustomDeletedFlag<D> = RxLocalDocumentData<D> & {
+    [k in typeof RXDB_POUCH_DELETED_FLAG]?: boolean;
+};
+
+export function localDocumentToPouch<D>(
+    docData: RxLocalDocumentData<D>
+): RxLocalDocumentDataWithCustomDeletedFlag<D> {
+    const ret: RxLocalDocumentDataWithCustomDeletedFlag<D> = flatClone(docData);
+
+    // add local prefix
+    ret._id = POUCHDB_LOCAL_PREFIX + ret._id;
+
+    // add custom deleted flag if document is deleted 
+    if (docData._deleted) {
+        ret._deleted = false;
+        ret[RXDB_POUCH_DELETED_FLAG] = true;
+    }
+
+    return ret;
+}
+
+export function localDocumentFromPouch<D>(
+    docData: RxLocalDocumentDataWithCustomDeletedFlag<D>
+): RxLocalDocumentData<D> {
+    const ret: RxLocalDocumentData<D> = flatClone(docData);
+
+    // strip local prefix
+    ret._id = ret._id.slice(POUCHDB_LOCAL_PREFIX_LENGTH);
+
+    if (docData[RXDB_POUCH_DELETED_FLAG]) {
+        ret._deleted = true;
+        delete (ret as any)[RXDB_POUCH_DELETED_FLAG];
+    }
+
+
+    return ret;
 }