+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Unified the replication primitives and the GraphQL replication plugin.
 - Removed the deprecated in-memory plugin.
 - Refactor usage of RxCollection.storageInstance to ensure all hooks run properly.
+- Added `withDeleted` parameter to `RxStorageKeyObjectInstance.findLocalDocumentsById()`
 
 ### 11.6.0 (4 February 2022)